add the clear button, split out UuidEntry and some :lipstick:
