+++ src/render/js/components/GenerateUuid.js
@@ -1,12 +1,7 @@
 import m from 'mithril'
-import { ipcRenderer } from 'electron'
 
-import sendToChannel from 'shared/sendToChannel'
-
-import UuidItem from './UuidItem'
-
-const sendTask    = sendToChannel(ipcRenderer, 'task')
-const requestUUID = (fn, prop) => () => fn({ type: 'uuid', num: prop() })
+import UuidItem   from './UuidItem'
+import UuidEntry  from './UuidEntry'
 
 function mapItems(Comp, attrs) {
   const { dispatch } = attrs
@@ -16,37 +11,13 @@ function mapItems(Comp, attrs) {
   }
 }
 
-function controller(attrs) {
-  const num     = m.prop('')
-  const request = requestUUID(sendTask, num)
-
-  return { num, request }
-}
-
 function view(ctrl, attrs) {
-  const { num, request } = ctrl
   const { uuids, dispatch } = attrs
   const results = uuids || []
 
   return (
     <div className="uuid">
-      <div className="uuid__entry form--inline">
-        <label className="label">
-          <span className="label__text">Number:</span>
-          <input
-            className="text"
-            size="3"
-            maxlength="3"
-            onchange={m.withAttr('value', num)}
-            value={num()}
-          />
-        </label>
-        <button
-          className="button uuid__button"
-          type="button"
-          onclick={request}
-        >Generate</button>
-      </div>
+      <UuidEntry dispatch={dispatch} />
       <ul className="uuid__results">
         {results.map(mapItems(UuidItem, { dispatch }))}
       </ul>
@@ -54,6 +25,6 @@ function view(ctrl, attrs) {
   )
 }
 
-const GenerateUUID = { view, controller }
+const GenerateUUID = { view }
 
 export default GenerateUUID
