+++ src/render/js/store/actions/uuid.js
@@ -1,5 +1,6 @@
-export const UUID_ADD       = 'UUID_ADD'
-export const UUID_DELETE    = 'UUID_DELETE'
+export const UUID_ADD     = 'UUID_ADD'
+export const UUID_DELETE  = 'UUID_DELETE'
+export const UUID_CLEAR   = 'UUID_CLEAR'
 
 export function addUuid(uuids) {
   return {
@@ -12,5 +13,11 @@ export function deleteUuid(index) {
   return {
     type: UUID_DELETE,
     index
+  }
+}
+
+export function clearUuid() {
+  return {
+    type: UUID_CLEAR
   }
 }
